Inititalize render implementation
