+++ src/js/search.js
@@ -1,11 +1,12 @@
 import MovieApiService from './api-service';
 import Notification from './notifications';
-import movieCardTpl from '../templates/movie-card.hbs';
+// import movieCardTpl from '../templates/movie-card.hbs';
+import { processResponse, renderMovies } from './render';
 
 import 'animate.css';
 
 const formRef = document.querySelector('#search-form');
-const galleryRef = document.querySelector('#gallery');
+// const galleryRef = document.querySelector('#gallery');
 
 formRef.addEventListener('submit', onSubmit);
 
@@ -44,37 +45,37 @@ async function searchMovies(searchQuery, page) {
   }
 }
 
-function processResponse(movies) {
-  if (movies.results.length === 0) {
-    notify.notFound();
-  }
-  const IMAGE_BASE_URL = localStorage.getItem('img_base_url');
-  const genresList = JSON.parse(localStorage.getItem('genres')).genres;
+// function processResponse(movies) {
+//   if (movies.results.length === 0) {
+//     notify.notFound();
+//   }
+//   const IMAGE_BASE_URL = localStorage.getItem('img_base_url');
+//   const genresList = JSON.parse(localStorage.getItem('genres')).genres;
 
-  const moviesProcessed = movies.results.map(
-    ({ id, release_date, title, poster_path, genre_ids }) => {
-      const genresNamed = genresList
-        .filter(genre => genre_ids.includes(genre.id))
-        .map(genre => genre.name);
+//   const moviesProcessed = movies.results.map(
+//     ({ id, release_date, title, poster_path, genre_ids }) => {
+//       const genresNamed = genresList
+//         .filter(genre => genre_ids.includes(genre.id))
+//         .map(genre => genre.name);
 
-      return {
-        id,
-        release_date: release_date ? release_date.slice(0, 4) : 'Date unknown',
-        title,
-        posterURL: poster_path ? `${IMAGE_BASE_URL}w500${poster_path}` : '',
-        genres:
-          genresNamed.length > 2
-            ? genresNamed.slice(0, 2).concat('Other').join(', ')
-            : genresNamed.join(', '),
-      };
-    },
-  );
+//       return {
+//         id,
+//         release_date: release_date ? release_date.slice(0, 4) : 'Date unknown',
+//         title,
+//         posterURL: poster_path ? `${IMAGE_BASE_URL}w500${poster_path}` : '',
+//         genres:
+//           genresNamed.length > 2
+//             ? genresNamed.slice(0, 2).concat('Other').join(', ')
+//             : genresNamed.join(', '),
+//       };
+//     },
+//   );
 
-  return moviesProcessed;
-}
+//   return moviesProcessed;
+// }
 
-function renderMovies(movies) {
-  galleryRef.innerHTML = movieCardTpl(movies);
-}
+// function renderMovies(movies) {
+//   galleryRef.innerHTML = movieCardTpl(movies);
+// }
 
-export { searchMovies, processResponse, renderMovies };
+// export { searchMovies, processResponse, renderMovies };
